finished part 1 of lesson 1.2
